requirements.txt
fastapi==0.95.1
asyncpraw==7.7.0
prawcore==2.3.0
python-dotenv==0.11.0
requests==2.28.2
sentry-sdk==1.21.0
uvicorn==0.22.0
